src/requirements.txt
-i https://pypi.org/simple
​
altair==4.2.0 ; python_version >= '3.7'
​
attrs==22.2.0 ; python_version >= '3.6'
​
blinker==1.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
​
blis==0.7.9
​
cachetools==5.2.0 ; python_version ~= '3.7'
​
catalogue==2.0.8 ; python_version >= '3.6'
​
certifi==2022.12.7 ; python_version >= '3.6'
​
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
​
click==8.1.3 ; python_version >= '3.7'
​
colorama==0.4.6 ; platform_system == 'Windows'
​
commonmark==0.9.1
​
confection==0.0.3 ; python_version >= '3.6'
​
cymem==2.0.7
​
decorator==5.1.1 ; python_version >= '3.5'
​
entrypoints==0.4 ; python_version >= '3.6'
​
gitdb==4.0.10 ; python_version >= '3.7'
​
gitpython==3.1.29 ; python_version >= '3.7'
​
idna==3.4 ; python_version >= '3.5'
​
importlib-metadata==5.2.0 ; python_version >= '3.7'
​
jinja2==3.1.2 ; python_version >= '3.7'
​
joblib==1.2.0 ; python_version >= '3.7'
​
jsonschema==4.17.3 ; python_version >= '3.7'
​
langcodes==3.3.0 ; python_version >= '3.6'
​
markupsafe==2.1.1 ; python_version >= '3.7'
​
murmurhash==1.0.9 ; python_version >= '3.6'
​
numpy==1.24.1
​
packaging==22.0 ; python_version >= '3.7'
​
pandas==1.5.2
​
pathy==0.10.1 ; python_version >= '3.6'
​
pillow==9.3.0 ; python_version >= '3.7'
​
preshed==3.0.8 ; python_version >= '3.6'
​
protobuf==3.20.3 ; python_version >= '3.7'
​
pyarrow==10.0.1 ; python_version >= '3.7'
​
pydantic==1.10.2 ; python_version >= '3.7'
​
pydeck==0.8.0 ; python_version >= '3.7'
​
pygments==2.13.0 ; python_version >= '3.6'
​
pympler==1.0.1 ; python_version >= '3.6'
​
pyrsistent==0.19.2 ; python_version >= '3.7'
​
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
​
pytz==2022.7
​
pytz-deprecation-shim==0.1.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
​
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
​
rich==12.6.0 ; python_full_version >= '3.6.3' and python_full_version < '4.0.0'
​
scikit-learn==1.2.0
​
scipy==1.9.3 ; python_version >= '3.8'
​
semver==2.13.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
​
setuptools==65.6.3 ; python_version >= '3.7'
​
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
​
smart-open==6.3.0 ; python_version >= '3.6' and python_version < '4.0'
​
smmap==5.0.0 ; python_version >= '3.6'
​
spacy==3.4.4
​
spacy-legacy==3.0.10 ; python_version >= '3.6'
​
spacy-loggers==1.0.4 ; python_version >= '3.6'
​
srsly==2.4.5 ; python_version >= '3.6'
​
streamlit==1.16.0
​
thinc==8.1.6 ; python_version >= '3.6'
​
threadpoolctl==3.1.0 ; python_version >= '3.6'
​
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
​
toolz==0.12.0 ; python_version >= '3.5'
​
tornado==6.2 ; python_version >= '3.7'
​
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
​
typer==0.7.0 ; python_version >= '3.6'
​
typing-extensions==4.4.0 ; python_version >= '3.7'
​
tzdata==2022.7 ; platform_system == 'Windows'
​
tzlocal==4.2 ; python_version >= '3.6'
​
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
​
validators==0.20.0 ; python_version >= '3.4'
​
wasabi==0.10.1
​
watchdog==2.2.0 ; platform_system != 'Darwin'
​
zipp==3.11.0 ; python_version >= '3.7'